move cli to .mts extension
