+++ packages/cli/package.json
@@ -19,5 +19,6 @@
     "type-fest": "^2.16.0",
     "typesafe-utils": "^1.16.2",
     "typescript": "^4.7.4"
-  }
+  },
+  "type": "module"
 }
