+++ packages/cli/src/setup/package-json.mts
